+++ src/main/java/chess/domain/board/Board.java
@@ -1,10 +1,10 @@
 package chess.domain.board;
 
-import chess.domain.Color;
-import chess.domain.Position;
-import chess.domain.piece.maker.PiecesGenerator;
 import chess.domain.piece.BlankPiece;
 import chess.domain.piece.Piece;
+import chess.domain.piece.maker.PiecesGenerator;
+import chess.domain.piece.property.Color;
+import chess.domain.position.Position;
 
 import java.util.List;
 import java.util.Optional;
