+++ src/test/java/chess/domain/piece/property/ColorTest.java
@@ -1,4 +1,4 @@
-package chess.domain;
+package chess.domain.piece.property;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
