+++ src/test/java/chess/domain/state/ChessEndTest.java
@@ -3,12 +3,12 @@ package chess.domain.state;
 import chess.TestPiecesGenerator;
 import chess.controller.command.Command;
 import chess.domain.ChessGame;
-import chess.domain.Color;
-import chess.domain.File;
-import chess.domain.Position;
-import chess.domain.Rank;
 import chess.domain.piece.Pawn;
 import chess.domain.piece.Piece;
+import chess.domain.piece.property.Color;
+import chess.domain.position.File;
+import chess.domain.position.Position;
+import chess.domain.position.Rank;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
